Migrate LineStrips
