+++ rerun_py/rerun_sdk/rerun/_rerun2/components/_overrides/__init__.py
@@ -1,5 +1,3 @@
 from __future__ import annotations
 
-from .line_strip2d import line_strip2d__native_to_pa_array_override
-from .line_strip3d import line_strip3d__native_to_pa_array_override
 from .radius import radius__native_to_pa_array_override
